requirements.txt
requests
pinecone-client
python-dotenv
llama-index
llama-index-vector-stores-pinecone
llama-index-embeddings-openai
tiktoken

nltk
scipy

langchain
langchain_community
langchain_openai
# llama_index.llms.langchain

#crawler
firecrawl
crawl4ai
playwright
nest_asyncio
html2text

#chunking
langchain_experimental

#backend
fastapi
uvicorn

#frontend
streamlit
streamlit_chat

#validation
pydantic
typing

#langauge
langdetect
googletrans
langchain_google_genai

langfuse

gspread
oauth2client